Extract note-modal subsystem into four sections/ modules
Completes the dashboard-main.js region split. note-preview.js owns _noteModal + _noteNavStack and the open/navigate/close flow; note-mode.js owns view/cm/plain switching, saveEdit, createNoteFor, startRenameNote, and the beforeunload dirty-guard; in-note-search.js owns the Ctrl+F / Ctrl+E capture-phase keydown + the search bar; note-reconcile.js owns the Phase 7 external-change reconcile with its own reconcileState object.

reconcileState (suppressedUntilMtime) follows the termState / reloadState / staleState / sseState pattern so preview's open/close can reset it across the ESM live-binding limit. _noteModal isn't an export rebind target (only its fields are mutated) so plain export works there.

Deleted: toggleNotes (12 LOC) — confirmed no callers after a grep across .js/.py/.j2/.html/.css in the repo.

reload-guards.js, cm6-mount.js, sse.js retargeted to import _noteModal / saveEdit / _reconcileNoteModal from the new section modules instead of dashboard-main.js.

dashboard-main.js: 2416 -> 1547 LOC (-869). Start-to-date: 4484 -> 1547 LOC (-65.5%) across 20 section modules.

Design: projects/2026-04-23-condash-frontend-extraction/notes/06-p09-cut4.md
