+++ frontend/src/js/dashboard-main.js
@@ -75,6 +75,24 @@ import {
 import {
     initSseSideEffects,
 } from './sections/sse.js';
+import {
+    openNotePreview, noteNavBack, closeNotePreview,
+} from './sections/note-preview.js';
+import {
+    noteSearchRun, noteSearchStep, noteSearchClose,
+    initInNoteSearchSideEffects,
+} from './sections/in-note-search.js';
+import {
+    setNoteMode, saveEdit, createNoteFor, startRenameNote,
+    initNoteModeSideEffects,
+} from './sections/note-mode.js';
+import {
+    _reconcileNoteModal,
+    _noteReconcileDismiss, _noteReconcileReload,
+} from './sections/note-reconcile.js';
+import {
+    _setDirty,
+} from './sections/note-preview.js';
 
 /* --- In-app config editor --- */
 function _setField(form, name, value) {
@@ -1104,525 +1122,6 @@ function toggleSection(el) {
     }
 }
 
-/* --- Notes block (collapsible list) --- */
-function toggleNotes(el) {
-    var list = el.nextElementSibling;
-    if (list.style.display === 'none') {
-        list.style.display = 'block';
-        el.classList.add('open');
-    } else {
-        list.style.display = 'none';
-        el.classList.remove('open');
-    }
-}
-
-/* --- Note preview modal --- */
-function _renderMermaidIn(container) {
-    if (!window.mermaid) return;
-    var blocks = container.querySelectorAll('pre.mermaid, pre > code.language-mermaid');
-    if (!blocks.length) return;
-    var nodes = [];
-    blocks.forEach(function(block) {
-        var pre = block.tagName === 'PRE' ? block : block.parentElement;
-        var code = pre.querySelector('code') || pre;
-        var src = code.textContent;
-        var div = document.createElement('div');
-        div.className = 'mermaid';
-        div.textContent = src;
-        pre.replaceWith(div);
-        nodes.push(div);
-    });
-    var isDark = document.documentElement.getAttribute('data-theme') === 'dark';
-    try {
-        window.mermaid.initialize({
-            startOnLoad: false,
-            theme: isDark ? 'dark' : 'default',
-            securityLevel: 'strict',
-        });
-        window.mermaid.run({ nodes: nodes }).catch(function() {});
-    } catch (e) {}
-}
-
-/* Mount the vendored PDF.js viewer on every .note-pdf-host inside the
-   view pane. If the PDF.js ES module (at the bottom of this file) hasn't
-   resolved yet, mark hosts pending — the module's ready-hook will flush
-   them. Safe to call repeatedly; mount() no-ops on already-mounted hosts. */
-function _mountPdfsIn(container) {
-    if (!container) return;
-    var hosts = container.querySelectorAll('.note-pdf-host');
-    for (var i = 0; i < hosts.length; i++) {
-        var host = hosts[i];
-        if (host.dataset.mounted === '1') continue;
-        if (window.__pdfjs && window.__pdfjs.ready) {
-            window.__pdfjs.mount(host);
-        } else if (window.__pdfjs && window.__pdfjs.error) {
-            host.innerHTML = '<div class="pdf-error">PDF viewer failed to load.</div>';
-        } else {
-            host.dataset.pdfPending = '1';
-            host.innerHTML = '<div class="pdf-loading">Loading PDF viewer\u2026</div>';
-        }
-    }
-}
-
-/* Back-navigation stack. Every time a link inside the note modal opens
-   a different note (wikilinks, relative .md links), the currently shown
-   {path, name} is pushed before the replacement. The back button in the
-   modal header pops one level. Cleared on close — this is an in-modal
-   navigation history, not a persistent browser history. */
-var _noteNavStack = [];
-
-/* Notes open in three modes. The modal carries `data-mode` on its inner
-   element (#note-modal-inner); three sibling panes inside #note-modal-body
-   are hidden/shown by CSS. `_noteModal` tracks the state shared across
-   panes so mode switches preserve user edits and mtime for the save
-   contract. */
-export var _noteModal = {
-    path: null,
-    editable: false,     // false when kind is pdf/image/binary — edit modes disabled
-    kind: null,          // from /note-raw
-    mtime: null,
-    renderedHtml: '',    // last server render shown in the view pane
-    /* Canonical text shared between CM6 and the plain textarea. Updated
-       whenever the user switches away from an edit mode so the other
-       mode can start from the same buffer. Reset on open and on save. */
-    text: '',
-    /* Which edit mode was last active. Ctrl-E from view returns here. */
-    lastEditMode: 'cm',
-    /* Unsaved-changes flag. Set on every CM6/textarea edit, cleared on
-       open and successful save. Drives the Save button's disabled state
-       and the close/beforeunload confirms. */
-    dirty: false,
-};
-
-/* Flip the dirty flag and refresh the Save button. Safe to call on every
-   keystroke — the button toggle is the only DOM work. Also drains any
-   reload requests that were parked because the modal was dirty. */
-export function _setDirty(value) {
-    var next = !!value;
-    if (_noteModal.dirty === next) return;
-    _noteModal.dirty = next;
-    _syncSaveButton();
-    if (!next && typeof _flushPendingReloads === 'function') {
-        _flushPendingReloads();
-    }
-}
-
-/* Save button is enabled only when there are unsaved edits, so after a
-   successful save the user gets a clear "saved" signal and can't click
-   again redundantly. Disabled when clean, non-editable, or viewing. */
-function _syncSaveButton() {
-    var btn = document.getElementById('note-save-btn');
-    if (!btn) return;
-    var inner = document.getElementById('note-modal-inner');
-    var mode = inner ? inner.getAttribute('data-mode') : 'view';
-    var editing = mode === 'cm' || mode === 'plain';
-    btn.disabled = !editing || !_noteModal.editable || !_noteModal.dirty;
-    btn.title = btn.disabled && editing && _noteModal.editable
-        ? 'No unsaved changes' : 'Save (Ctrl+S)';
-}
-
-async function openNotePreview(path, name) {
-    var modal = document.getElementById('note-modal');
-    var inner = document.getElementById('note-modal-inner');
-    var title = document.getElementById('note-modal-title');
-    var viewPane = document.getElementById('note-pane-view');
-    var ta = document.getElementById('note-edit-textarea');
-    // Reset any in-note search: matches point at DOM nodes we're about
-    // to discard and the count would be stale against the new note.
-    noteSearchClose();
-    _destroyCm();
-    title.textContent = name;
-    _noteModal.path = path;
-    _noteModal.editable = false;
-    _noteModal.kind = null;
-    _noteModal.mtime = null;
-    _noteModal.text = '';
-    _noteModal.renderedHtml = '';
-    _noteModal.dirty = false;
-    _noteShowExternalBanner(false);
-    _noteReconcileSuppressedUntilMtime = null;
-    ta.value = '';
-    viewPane.innerHTML = '<p class="note-loading">Loading\u2026</p>';
-    _setNoteModeAttr(inner, 'view');
-    _syncModeControls();
-    _hideSaveError();
-    modal.classList.add('open');
-    // Kick off both fetches in parallel. /note returns HTML for the view
-    // pane; /note-raw returns text+mtime+kind for the edit panes. The
-    // raw fetch is best-effort: pdf/image kinds 4xx here and we leave
-    // the edit modes disabled.
-    var viewP = fetch('/note?path=' + encodeURIComponent(path)).then(function(res) {
-        if (!res.ok) throw new Error('HTTP ' + res.status);
-        return res.text();
-    });
-    var rawP = fetch('/note-raw?path=' + encodeURIComponent(path)).then(function(res) {
-        if (!res.ok) return null;
-        return res.json();
-    }).catch(function() { return null; });
-    try {
-        var html = await viewP;
-        // Guard against late arrival for a different note (user opened
-        // another before this one resolved).
-        if (_noteModal.path !== path) return;
-        viewPane.innerHTML = html;
-        viewPane.scrollTop = 0;
-        _noteModal.renderedHtml = html;
-        _renderMermaidIn(viewPane);
-        _wireNoteLinks(viewPane, path);
-        _mountPdfsIn(viewPane);
-    } catch (e) {
-        if (_noteModal.path !== path) return;
-        viewPane.innerHTML = '<p class="note-error">Failed to load note.</p>';
-    }
-    var raw = await rawP;
-    if (_noteModal.path !== path) return;
-    if (raw && typeof raw.content === 'string') {
-        _noteModal.editable = true;
-        _noteModal.kind = raw.kind || null;
-        _noteModal.mtime = raw.mtime != null ? Number(raw.mtime) : null;
-        _noteModal.text = raw.content;
-        ta.value = raw.content;
-    }
-    _syncModeControls();
-    _syncNoteBack();
-}
-
-/* Push the currently-shown note onto the back stack and open the target.
-   Used by link handlers inside the note modal (wikilinks + internal .md).
-   The button-driven "back" path calls openNotePreview without pushing so
-   the stack only grows on forward navigation. */
-async function _navigateToNote(path, name, anchor) {
-    if (_noteModal.path) {
-        var titleEl = document.getElementById('note-modal-title');
-        var currentName = titleEl ? titleEl.textContent : _noteModal.path;
-        _noteNavStack.push({path: _noteModal.path, name: currentName});
-    }
-    await openNotePreview(path, name);
-    if (anchor) _scrollNoteToAnchor(anchor);
-}
-
-function _syncNoteBack() {
-    var btn = document.getElementById('note-modal-back');
-    if (!btn) return;
-    if (_noteNavStack.length > 0) btn.removeAttribute('hidden');
-    else btn.setAttribute('hidden', '');
-}
-
-async function noteNavBack() {
-    if (_noteNavStack.length === 0) return;
-    var entry = _noteNavStack.pop();
-    await openNotePreview(entry.path, entry.name);
-}
-
-/* Resolve a path relative to `baseDir` — `baseDir/../foo` collapses to
-   `foo`, absolute paths drop the leading slash so they are treated as
-   conception-tree relative. */
-function _resolveNotePath(baseDir, rel) {
-    if (rel.startsWith('/')) rel = rel.replace(/^\/+/, '');
-    var parts = ((baseDir ? baseDir + '/' : '') + rel).split('/');
-    var out = [];
-    for (var i = 0; i < parts.length; i++) {
-        var p = parts[i];
-        if (!p || p === '.') continue;
-        if (p === '..') out.pop();
-        else out.push(p);
-    }
-    return out.join('/');
-}
-
-/* Scroll a pandoc-generated heading into view inside the modal pane.
-   Heading ids are produced by pandoc's gfm auto_identifiers; the caller
-   supplies the fragment from a link href. No-op if the id is missing. */
-function _scrollNoteToAnchor(anchor) {
-    if (!anchor) return;
-    var pane = document.getElementById('note-pane-view');
-    if (!pane) return;
-    var el = null;
-    try { el = pane.querySelector('#' + CSS.escape(anchor)); } catch (_) {}
-    if (el) el.scrollIntoView({block: 'start'});
-}
-
-/* Route note-body link clicks:
-   - http(s) → POST /open-external → host browser (bypasses pywebview).
-   - in-page anchors (#foo), mailto: → leave default behaviour.
-   - relative .md (with optional #anchor) → resolve inside the conception
-     tree and open in the same modal via openNotePreview.
-   - anything else → resolve against the note's directory and POST /open-doc
-     so the OS default viewer handles PDFs, images, and other files. */
-function _wireNoteLinks(body, notePath) {
-    var noteDir = notePath.lastIndexOf('/') >= 0
-        ? notePath.substring(0, notePath.lastIndexOf('/'))
-        : '';
-    // Wikilinks resolved server-side carry an absolute (conception-tree
-    // relative) href. They open inside the modal, not via xdg-open.
-    body.querySelectorAll('a.wikilink[href]').forEach(function(a) {
-        a.addEventListener('click', function(ev) {
-            ev.preventDefault();
-            var href = a.getAttribute('href');
-            var label = a.textContent || href;
-            _navigateToNote(href, label);
-        });
-    });
-    // Unresolved wikilinks: click does nothing but flash the hover title.
-    body.querySelectorAll('a.wikilink-missing').forEach(function(a) {
-        a.addEventListener('click', function(ev) { ev.preventDefault(); });
-    });
-    body.querySelectorAll('a[href]:not(.wikilink):not(.wikilink-missing)').forEach(function(a) {
-        var href = a.getAttribute('href');
-        if (!href) return;
-        if (/^https?:\/\//i.test(href)) {
-            a.addEventListener('click', function(ev) {
-                ev.preventDefault();
-                fetch('/open-external', {
-                    method: 'POST',
-                    headers: {'Content-Type': 'application/json'},
-                    body: JSON.stringify({url: href}),
-                });
-            });
-            return;
-        }
-        if (href.startsWith('#') || href.startsWith('mailto:')) {
-            return;
-        }
-        var hashIdx = href.indexOf('#');
-        var pathPart = hashIdx >= 0 ? href.substring(0, hashIdx) : href;
-        var anchor = hashIdx >= 0 ? href.substring(hashIdx + 1) : '';
-        if (pathPart && /\.md$/i.test(pathPart)) {
-            var resolvedMd = _resolveNotePath(noteDir, pathPart);
-            a.addEventListener('click', function(ev) {
-                ev.preventDefault();
-                var label = a.textContent || resolvedMd;
-                _navigateToNote(resolvedMd, label, anchor);
-            });
-            return;
-        }
-        var resolved = _resolveNotePath(noteDir, pathPart || href);
-        a.addEventListener('click', function(ev) {
-            ev.preventDefault();
-            fetch('/open-doc', {
-                method: 'POST',
-                headers: {'Content-Type': 'application/json'},
-                body: JSON.stringify({path: resolved}),
-            });
-        });
-    });
-}
-
-function closeNotePreview() {
-    // Capture the active buffer so dirty reflects the latest keystrokes
-    // even if the user never switched modes before closing.
-    if (_noteModal.editable) _captureActiveBuffer();
-    if (_noteModal.dirty) {
-        if (!confirm('You have unsaved changes. Discard them?')) return;
-    }
-    document.getElementById('note-modal').classList.remove('open');
-    _destroyCm();
-    _hideSaveError();
-    _noteModal.path = null;
-    _noteModal.dirty = false;
-    _noteNavStack = [];
-    _syncNoteBack();
-    noteSearchClose();
-    _noteShowExternalBanner(false);
-    _noteReconcileSuppressedUntilMtime = null;
-    if (typeof _flushPendingReloads === 'function') _flushPendingReloads();
-}
-
-/* --- In-note search (Ctrl+F inside the note modal) ---
-   Walks .note-modal-body's text nodes, wraps each case-insensitive
-   substring match in <mark class="note-match">, and lets the user step
-   through them with Enter / Shift+Enter / F3. Scoped to the preview
-   view — Ctrl+F inside the edit textarea falls through so the browser's
-   native behaviour still works there. */
-var _noteSearch = { matches: [], idx: -1 };
-function _clearNoteMarks() {
-    var pane = document.getElementById('note-pane-view');
-    if (!pane) return;
-    pane.querySelectorAll('mark.note-match').forEach(function(m) {
-        var parent = m.parentNode;
-        while (m.firstChild) parent.insertBefore(m.firstChild, m);
-        parent.removeChild(m);
-    });
-    pane.normalize();
-}
-/* Return the PDF find API attached to a mounted .note-pdf-host inside
-   the open note modal, or null. The PDF viewer's mount() exposes this
-   at host.__pdfFind so the shared note-search-bar can drive it. */
-function _notePdfFind() {
-    var modal = document.getElementById('note-modal');
-    if (!modal || !modal.classList.contains('open')) return null;
-    var host = modal.querySelector('.note-pdf-host[data-mounted="1"]');
-    return host && host.__pdfFind ? host.__pdfFind : null;
-}
-function _setSearchCount(state, q) {
-    var countEl = document.getElementById('note-search-count');
-    if (!countEl) return;
-    if (!state || !state.matches.length) {
-        countEl.textContent = q ? '0/0' : '';
-    } else {
-        countEl.textContent = (state.idx + 1) + '/' + state.matches.length;
-    }
-}
-function noteSearchRun() {
-    var input = document.getElementById('note-search-input');
-    var q = input ? input.value : '';
-    var countEl = document.getElementById('note-search-count');
-    var pdfFind = _notePdfFind();
-    if (pdfFind) {
-        // Clear any stale view-pane marks and delegate to the PDF viewer.
-        _clearNoteMarks();
-        _noteSearch.matches = [];
-        _noteSearch.idx = -1;
-        pdfFind.run(q).then(function(state) { _setSearchCount(state, q); });
-        return;
-    }
-    _clearNoteMarks();
-    _noteSearch.matches = [];
-    _noteSearch.idx = -1;
-    if (!q) {
-        if (countEl) countEl.textContent = '';
-        return;
-    }
-    var pane = document.getElementById('note-pane-view');
-    if (!pane) return;
-    var qLow = q.toLowerCase();
-    var qLen = q.length;
-    // Collect text nodes first (the walker becomes unreliable once we
-    // mutate the tree). Skip script/style just in case; the body is
-    // server-rendered markdown but we don't assume.
-    var walker = document.createTreeWalker(pane, NodeFilter.SHOW_TEXT, {
-        acceptNode: function(n) {
-            var tag = n.parentNode && n.parentNode.nodeName;
-            if (tag === 'SCRIPT' || tag === 'STYLE') return NodeFilter.FILTER_REJECT;
-            return NodeFilter.FILTER_ACCEPT;
-        }
-    });
-    var textNodes = [];
-    var node;
-    while ((node = walker.nextNode())) textNodes.push(node);
-    textNodes.forEach(function(n) {
-        var low = n.nodeValue.toLowerCase();
-        // Iterate in reverse so splitText offsets earlier in the string
-        // stay valid as we carve the node from right to left.
-        var positions = [];
-        var pos = 0;
-        while ((pos = low.indexOf(qLow, pos)) !== -1) {
-            positions.push(pos);
-            pos += qLen;
-        }
-        for (var i = positions.length - 1; i >= 0; i--) {
-            var start = positions[i];
-            var matchNode = n.splitText(start);
-            matchNode.splitText(qLen);
-            var mark = document.createElement('mark');
-            mark.className = 'note-match';
-            matchNode.parentNode.replaceChild(mark, matchNode);
-            mark.appendChild(matchNode);
-            _noteSearch.matches.push(mark);
-        }
-    });
-    // Sort matches into document order (per-node reverse iteration left
-    // them reverse within each node, and textNodes was walked forward).
-    _noteSearch.matches.sort(function(a, b) {
-        var cmp = a.compareDocumentPosition(b);
-        if (cmp & Node.DOCUMENT_POSITION_FOLLOWING) return -1;
-        if (cmp & Node.DOCUMENT_POSITION_PRECEDING) return 1;
-        return 0;
-    });
-    if (_noteSearch.matches.length) {
-        _noteSearch.idx = 0;
-        _noteSearch.matches[0].classList.add('active');
-        _noteSearch.matches[0].scrollIntoView({block: 'center'});
-    }
-    if (countEl) {
-        countEl.textContent = _noteSearch.matches.length
-            ? (_noteSearch.idx + 1) + '/' + _noteSearch.matches.length
-            : '0/0';
-    }
-}
-function noteSearchStep(dir) {
-    var pdfFind = _notePdfFind();
-    if (pdfFind) {
-        var state = pdfFind.step(dir);
-        _setSearchCount(state, '');
-        return;
-    }
-    var n = _noteSearch.matches.length;
-    if (!n) return;
-    if (_noteSearch.idx >= 0) {
-        _noteSearch.matches[_noteSearch.idx].classList.remove('active');
-    }
-    _noteSearch.idx = (_noteSearch.idx + dir + n) % n;
-    var m = _noteSearch.matches[_noteSearch.idx];
-    m.classList.add('active');
-    m.scrollIntoView({block: 'center'});
-    var countEl = document.getElementById('note-search-count');
-    if (countEl) countEl.textContent = (_noteSearch.idx + 1) + '/' + n;
-}
-function noteSearchOpen() {
-    var bar = document.getElementById('note-search-bar');
-    if (!bar) return;
-    bar.hidden = false;
-    var input = document.getElementById('note-search-input');
-    if (input) { input.focus(); input.select(); }
-    // If a query is already typed, re-run so marks come back after the
-    // note was reloaded or the bar was reopened.
-    if (input && input.value) noteSearchRun();
-}
-function noteSearchClose() {
-    var pdfFind = _notePdfFind();
-    if (pdfFind) pdfFind.close();
-    var bar = document.getElementById('note-search-bar');
-    if (bar) bar.hidden = true;
-    _clearNoteMarks();
-    _noteSearch.matches = [];
-    _noteSearch.idx = -1;
-    var input = document.getElementById('note-search-input');
-    if (input) input.value = '';
-    var countEl = document.getElementById('note-search-count');
-    if (countEl) countEl.textContent = '';
-}
-/* Capture-phase keydown so this beats the existing Escape handler and
-   any xterm/editor shortcuts that might swallow Ctrl+F elsewhere. */
-document.addEventListener('keydown', function(ev) {
-    var modal = document.getElementById('note-modal');
-    if (!modal || !modal.classList.contains('open')) return;
-    var inner = document.getElementById('note-modal-inner');
-    var mode = inner ? inner.getAttribute('data-mode') : 'view';
-    var editing = mode === 'cm' || mode === 'plain';
-    var isFindKey = (ev.ctrlKey || ev.metaKey) && !ev.altKey
-        && (ev.key === 'f' || ev.key === 'F');
-    if (isFindKey) {
-        if (editing) return;  // let the edit panes keep native behaviour
-        ev.preventDefault();
-        ev.stopPropagation();
-        noteSearchOpen();
-        return;
-    }
-    // Ctrl+E toggles between view and the last-used edit mode.
-    if ((ev.ctrlKey || ev.metaKey) && !ev.altKey && (ev.key === 'e' || ev.key === 'E')) {
-        if (!_noteModal.editable) return;
-        ev.preventDefault();
-        ev.stopPropagation();
-        setNoteMode(mode === 'view' ? (_noteModal.lastEditMode || 'cm') : 'view');
-        return;
-    }
-    var bar = document.getElementById('note-search-bar');
-    if (!bar || bar.hidden) return;
-    var activeInSearch = document.activeElement
-        && document.activeElement.id === 'note-search-input';
-    if (ev.key === 'Escape') {
-        ev.preventDefault();
-        ev.stopPropagation();
-        noteSearchClose();
-    } else if (ev.key === 'Enter' && activeInSearch) {
-        ev.preventDefault();
-        noteSearchStep(ev.shiftKey ? -1 : 1);
-    } else if (ev.key === 'F3') {
-        ev.preventDefault();
-        noteSearchStep(ev.shiftKey ? -1 : 1);
-    }
-}, true);
 
 /* Route a ## Deliverables PDF click to the OS default viewer. target="_blank"
    under pywebview routes to the system browser on 127.0.0.1:<port> and fails
@@ -1637,363 +1136,6 @@ function openDeliverable(path) {
     }).catch(function() {});
 }
 
-/* --- Note modal mode management ---
-   Three sibling panes (view, cm, plain) live inside #note-modal-body;
-   the modal's data-mode attribute drives CSS visibility. setNoteMode
-   transfers the canonical buffer between the two edit modes on switch
-   so edits survive toggling, then updates the chrome (save button,
-   toggle state). Save POSTs to /note with expected_mtime — server
-   refuses on mtime mismatch so a stale editor can't silently clobber
-   out-of-band edits. */
-
-function _setNoteModeAttr(inner, mode) {
-    inner.setAttribute('data-mode', mode);
-}
-
-function _hideSaveError() {
-    var err = document.getElementById('note-edit-error');
-    if (err) { err.textContent = ''; err.classList.remove('visible'); }
-}
-
-function _showSaveError(msg) {
-    var err = document.getElementById('note-edit-error');
-    if (!err) return;
-    err.textContent = msg;
-    err.classList.add('visible');
-}
-
-/* Update the mode toggle (disable edit buttons when the file is not
-   editable) and show/hide the Save button per mode. */
-function _syncModeControls() {
-    var inner = document.getElementById('note-modal-inner');
-    var mode = inner ? inner.getAttribute('data-mode') : 'view';
-    var toggle = document.getElementById('note-mode-toggle');
-    var saveBtn = document.getElementById('note-save-btn');
-    if (toggle) {
-        toggle.querySelector('[data-mode="cm"]').disabled =
-            !_noteModal.editable || !window.__cm6;
-        toggle.querySelector('[data-mode="plain"]').disabled = !_noteModal.editable;
-        if (!window.__cm6) {
-            toggle.querySelector('[data-mode="cm"]').title =
-                'Loading editor…';
-        } else if (!_noteModal.editable) {
-            toggle.querySelector('[data-mode="cm"]').title =
-                'This file is not editable (binary/preview-only)';
-        } else {
-            toggle.querySelector('[data-mode="cm"]').title =
-                'Edit with syntax highlighting (Ctrl+E)';
-        }
-    }
-    if (saveBtn) saveBtn.style.display = (mode === 'cm' || mode === 'plain') ? '' : 'none';
-    _syncSaveButton();
-}
-
-/* Pull the active pane's buffer back into _noteModal.text so a mode
-   switch starts the next pane from the same content. */
-function _captureActiveBuffer() {
-    var inner = document.getElementById('note-modal-inner');
-    var mode = inner.getAttribute('data-mode');
-    if (mode === 'cm' && _cm.view) {
-        _noteModal.text = _cm.view.state.doc.toString();
-    } else if (mode === 'plain') {
-        var ta = document.getElementById('note-edit-textarea');
-        if (ta) _noteModal.text = ta.value;
-    }
-}
-
-/* Hydrate a pane from _noteModal.text so the user sees their latest
-   edits after a switch. No-op for view mode — view reflects last-save. */
-function _hydratePane(mode) {
-    if (mode === 'plain') {
-        var ta = document.getElementById('note-edit-textarea');
-        if (!ta) return;
-        if (ta.value !== _noteModal.text) ta.value = _noteModal.text;
-        ta.setSelectionRange(0, 0);
-        ta.scrollTop = 0;
-    } else if (mode === 'cm') {
-        if (!_cm.view) { _mountCm(); return; }
-        var cur = _cm.view.state.doc.toString();
-        if (cur !== _noteModal.text) {
-            _cm.view.dispatch({
-                changes: {from: 0, to: cur.length, insert: _noteModal.text},
-            });
-        }
-        _cm.view.dispatch({selection: {anchor: 0}});
-        _cm.view.scrollDOM.scrollTop = 0;
-    } else if (mode === 'view') {
-        var pane = document.getElementById('note-pane-view');
-        if (pane) pane.scrollTop = 0;
-    }
-}
-
-function setNoteMode(next) {
-    if (next !== 'view' && next !== 'cm' && next !== 'plain') return;
-    if ((next === 'cm' || next === 'plain') && !_noteModal.editable) return;
-    if (next === 'cm' && !window.__cm6) return;
-    var inner = document.getElementById('note-modal-inner');
-    var prev = inner.getAttribute('data-mode');
-    if (prev === next) return;
-    if (prev === 'cm' || prev === 'plain') _captureActiveBuffer();
-    _setNoteModeAttr(inner, next);
-    if (next === 'cm' || next === 'plain') _noteModal.lastEditMode = next;
-    _hideSaveError();
-    _hydratePane(next);
-    _syncModeControls();
-    if (next === 'cm' && _cm.view) { setTimeout(function() { _cm.view.focus(); }, 0); }
-    else if (next === 'plain') {
-        var ta = document.getElementById('note-edit-textarea');
-        if (ta) setTimeout(function() { ta.focus(); }, 0);
-    }
-}
-
-export async function saveEdit() {
-    var inner = document.getElementById('note-modal-inner');
-    var mode = inner.getAttribute('data-mode');
-    if (mode !== 'cm' && mode !== 'plain') return;
-    if (!_noteModal.path) return;
-    _captureActiveBuffer();
-    _hideSaveError();
-    try {
-        var res = await fetch('/note', {
-            method: 'POST',
-            headers: {'Content-Type': 'application/json'},
-            body: JSON.stringify({
-                path: _noteModal.path,
-                content: _noteModal.text,
-                expected_mtime: _noteModal.mtime,
-            }),
-        });
-        var data = await res.json().catch(function() { return {}; });
-        if (!res.ok) {
-            if (res.status === 409 && data.mtime) _noteModal.mtime = Number(data.mtime);
-            _showSaveError(data.reason || data.error || ('HTTP ' + res.status));
-            return;
-        }
-        if (data.mtime != null) _noteModal.mtime = Number(data.mtime);
-        _setDirty(false);
-        // Refresh the view pane from the server render so it stays
-        // aligned with what's now on disk.
-        var name = document.getElementById('note-modal-title').textContent;
-        await _reloadNotePreview(_noteModal.path, name);
-    } catch (e) {
-        _showSaveError('Save failed: ' + e);
-    }
-}
-
-/* Prompt for a filename, POST /note/create, and drop into edit mode
-   on success. Default extension is `.md`; user can type something else. */
-/* Prompt for a filename and POST /note/create. ``subRelToNotes`` is the
-   target subdirectory relative to ``<item>/notes/`` ("" for notes/ root)
-   so the per-folder + buttons can drop the file into the folder whose
-   summary they live on. */
-async function createNoteFor(readmePath, subRelToNotes) {
-    var raw = prompt('New note filename (e.g. plan.md, decision.txt):', 'new-note.md');
-    if (!raw) return;
-    raw = raw.trim();
-    if (!raw) return;
-    if (raw.indexOf('.') < 0) raw = raw + '.md';
-    try {
-        var res = await fetch('/note/create', {
-            method: 'POST',
-            headers: {'Content-Type': 'application/json'},
-            body: JSON.stringify({
-                item_readme: readmePath,
-                filename: raw,
-                subdir: subRelToNotes || '',
-            }),
-        });
-        var data = await res.json().catch(function(){return {};});
-        if (!res.ok) {
-            alert('Could not create note: ' + (data.reason || data.error || ('HTTP ' + res.status)));
-            return;
-        }
-        await openNotePreview(data.path, raw);
-        // Drop straight into the preferred edit mode for the new file.
-        if (_noteModal.editable) {
-            setNoteMode(window.__cm6 ? 'cm' : 'plain');
-        }
-    } catch (e) {
-        alert('Network error: ' + e);
-    }
-}
-
-/* Double-click the modal title to rename the current note. Only files
-   under <item>/notes/** are renamable; READMEs and knowledge/* are
-   left alone (server returns 400 for those so the UI fails loud). The
-   extension stays fixed — users type a stem, server re-appends the
-   suffix. */
-var _NOTES_RENAMEABLE_RE = /^projects\/\d{4}-\d{2}\/\d{4}-\d{2}-\d{2}-[\w.\-]+\/notes\//;
-
-function startRenameNote() {
-    var titleEl = document.getElementById('note-modal-title');
-    var path = _noteModal.path || '';
-    if (!_NOTES_RENAMEABLE_RE.test(path)) return;
-    if (titleEl.querySelector('.note-rename-input')) return;  // already editing
-    var filename = path.substring(path.lastIndexOf('/') + 1);
-    var dotIdx = filename.lastIndexOf('.');
-    var stem = dotIdx > 0 ? filename.substring(0, dotIdx) : filename;
-    var ext = dotIdx > 0 ? filename.substring(dotIdx) : '';
-    var originalText = titleEl.textContent;
-    var restored = false;
-    var restore = function() {
-        if (restored) return;
-        restored = true;
-        titleEl.textContent = originalText;
-    };
-    var input = document.createElement('input');
-    input.type = 'text';
-    input.className = 'note-rename-input';
-    input.value = stem;
-    var extEl = document.createElement('span');
-    extEl.className = 'note-rename-ext';
-    extEl.textContent = ext;
-    titleEl.textContent = '';
-    titleEl.appendChild(input);
-    titleEl.appendChild(extEl);
-    input.focus();
-    input.select();
-    var commit = function() {
-        var newStem = input.value.trim();
-        if (!newStem || newStem === stem) { restore(); return; }
-        fetch('/note/rename', {
-            method: 'POST',
-            headers: {'Content-Type': 'application/json'},
-            body: JSON.stringify({path: path, new_stem: newStem}),
-        }).then(function(r) {
-            return r.json().then(function(data) { return {ok: r.ok, data: data}; });
-        }).then(function(result) {
-            if (!result.ok || !result.data.ok) {
-                alert('Rename failed: ' + (result.data.error || result.data.reason || 'unknown'));
-                restore();
-                return;
-            }
-            _noteModal.path = result.data.path;
-            if (result.data.mtime != null) _noteModal.mtime = Number(result.data.mtime);
-            var newName = newStem + ext;
-            titleEl.textContent = newName;
-            restored = true;  // keep the new value
-        }).catch(function(err) {
-            alert('Rename failed: ' + err);
-            restore();
-        });
-    };
-    input.onkeydown = function(ev) {
-        if (ev.key === 'Enter') { ev.preventDefault(); commit(); }
-        else if (ev.key === 'Escape') { ev.preventDefault(); restore(); }
-        ev.stopPropagation();
-    };
-    input.onblur = commit;
-    input.onclick = function(ev) { ev.stopPropagation(); };
-}
-
-/* --- Phase 7: note modal reconcile ---
-   When /events signals a filesystem change, any open note may now
-   diverge from its on-disk content. _reconcileNoteModal compares the
-   loaded mtime with the live mtime and acts by buffer state:
-     - clean buffer → silently reload text + render, restoring selection.
-     - dirty buffer → reveal the banner and leave the buffer alone until
-       the user picks "Keep my edits" or "Reload from disk".
-   The banner dismissal is sticky until a new external change is
-   detected (so the user who keeps editing isn't re-nagged).
-*/
-var _noteReconcileSuppressedUntilMtime = null;
-
-export async function _reconcileNoteModal() {
-    if (!_noteModal || !_noteModal.path) return;
-    var path = _noteModal.path;
-    try {
-        var res = await fetch('/note-raw?path=' + encodeURIComponent(path));
-        if (!res.ok) return;
-        var data = await res.json();
-        if (_noteModal.path !== path) return;  // user switched notes mid-flight
-        var fresh = Number(data.mtime);
-        var loaded = Number(_noteModal.mtime);
-        if (!isFinite(fresh) || !isFinite(loaded)) return;
-        if (fresh <= loaded) return;  // no external change
-        if (_noteReconcileSuppressedUntilMtime != null
-            && fresh <= _noteReconcileSuppressedUntilMtime) {
-            return;  // user said "keep my edits"; wait for a newer change
-        }
-        if (_noteModal.editable) _captureActiveBuffer();
-        if (_noteModal.dirty) {
-            _noteShowExternalBanner(true);
-        } else {
-            await _noteSilentReload(data);
-        }
-    } catch (e) {}
-}
-
-function _noteShowExternalBanner(show) {
-    var banner = document.getElementById('note-modal-external-banner');
-    if (!banner) return;
-    if (show) banner.removeAttribute('hidden');
-    else banner.setAttribute('hidden', '');
-}
-
-function _noteReconcileDismiss() {
-    // Remember the disk mtime we just decided to ignore so repeat
-    // polls don't re-show the banner until a newer change lands.
-    _noteReconcileSuppressedUntilMtime = Number(_noteModal.mtime) || 0;
-    _noteShowExternalBanner(false);
-}
-
-async function _noteReconcileReload() {
-    try {
-        var res = await fetch('/note-raw?path=' + encodeURIComponent(_noteModal.path));
-        if (!res.ok) return;
-        var data = await res.json();
-        await _noteSilentReload(data);
-        _noteShowExternalBanner(false);
-        _noteReconcileSuppressedUntilMtime = null;
-    } catch (e) {}
-}
-
-async function _noteSilentReload(rawData) {
-    // Update buffer + preview, preserve caret/selection where possible.
-    _noteModal.text = rawData.content;
-    _noteModal.mtime = Number(rawData.mtime);
-    _setDirty(false);
-    // Hydrate whichever pane is active. CM6 owns its own state —
-    // replace the doc but keep the selection when it still fits.
-    if (_cm && _cm.view) {
-        var prevSel = null;
-        try { prevSel = _cm.view.state.selection; } catch (e) {}
-        _cm.view.dispatch({
-            changes: {from: 0, to: _cm.view.state.doc.length, insert: _noteModal.text},
-        });
-        if (prevSel) {
-            try {
-                var max = _cm.view.state.doc.length;
-                var anchor = Math.min(prevSel.main.anchor, max);
-                var head = Math.min(prevSel.main.head, max);
-                _cm.view.dispatch({selection: {anchor: anchor, head: head}});
-            } catch (e) {}
-        }
-    }
-    var ta = document.getElementById('note-edit-textarea');
-    if (ta) ta.value = _noteModal.text;
-    try {
-        await _reloadNotePreview(_noteModal.path, null);
-    } catch (e) {}
-}
-
-async function _reloadNotePreview(path, name) {
-    var pane = document.getElementById('note-pane-view');
-    if (name) document.getElementById('note-modal-title').textContent = name;
-    pane.innerHTML = '<p class="note-loading">Loading\u2026</p>';
-    var res = await fetch('/note?path=' + encodeURIComponent(path));
-    if (!res.ok) {
-        pane.innerHTML = '<p class="note-error">Failed to load note (' + res.status + ').</p>';
-        return;
-    }
-    var html = await res.text();
-    pane.innerHTML = html;
-    _noteModal.renderedHtml = html;
-    _renderMermaidIn(pane);
-    _wireNoteLinks(pane, path);
-    _mountPdfsIn(pane);
-    pane.scrollTop = 0;
-}
 
 document.addEventListener('keydown', function(e) {
     if (e.key !== 'Escape') return;
@@ -2015,19 +1157,6 @@ document.addEventListener('keydown', function(e) {
     }
 });
 
-/* Guard window-level navigation (tab close, reload, external nav) while
-   a note modal has unsaved edits. The `returnValue` dance is required
-   for legacy cross-browser support; modern browsers ignore the string
-   and show their own generic confirm. */
-window.addEventListener('beforeunload', function(e) {
-    var modal = document.getElementById('note-modal');
-    if (!modal || !modal.classList.contains('open')) return;
-    if (_noteModal.editable) _captureActiveBuffer();
-    if (!_noteModal.dirty) return;
-    e.preventDefault();
-    e.returnValue = '';
-    return '';
-});
 
 async function cycle(file, line, el) {
     var res = await fetch('/toggle', {
@@ -2354,6 +1483,8 @@ initNotesTreeStateSideEffects();
 initCm6ThemeSyncSideEffects();
 initGitActionsSideEffects();
 initRunnerViewersSideEffects();
+initInNoteSearchSideEffects();
+initNoteModeSideEffects();
 
 // Phase 6: event-driven staleness. /events streams tab-level hints;
 // checkUpdates() runs on connect + every hint to reconcile the real
